requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
flask-socketio==5.3.6
eventlet==0.36.1
requests==2.32.3
python-dotenv==1.0.1
